Moved socket implementation to navbar
